Extracted controller from error console view to enable tests
